Allow user to specify reminder date and time
Currently, the .remindmein command actually does allow for specifying a
date and time, but the main issue was that it doesn't handle timezones
very well. The bot runs in UTC while most users will enter in times in
their own timezone.

As of now, a Vancouver user entering a command like ".remindmein 2:00pm"
will register a time 7 or 8 hours earlier than they likely intended.

This should allow users to enter in a time and date with an optional
timezone, defaulting to "Canada/Pacific" if none is given (I feel this
is fair given that most of SFU's students are in the Metro Vancouver
area).

Should resolve #216